Revert "Refactor to allow supporting multiple db types with clean code."
This reverts commit 54d7a78f672dd553c4c3a7ae19fd24ad5af7700b.
